Canonicalize mtree paths to ensure path comparison is exact
The `mtree` passed to the `tar` rule is user-provided. Unlike the other
streams participating in this comparison, we cannot completely trust
that the encoding it uses is the same as the form this ruleset would
produce. There's a bit of play to how libarchive interprets these path
specs. To be correct, the comparison we're making requires that all
equivalent forms be massaged to a single representation before
performing the checks.