Decode vis-encoded file paths when passing to Bazel
Bazel expects file paths to be written verbatim in the unused inputs
file; it does not understand any escaping or encoding scheme.

Because we're not post-filtering depsets written out from Bazel, we're
no longer able to produce a 2-column format as easily: sed performs
replacements in all columns. That's fine for the
intended-to-be-vis-encoded column 1, but incorrect for the
intended-to-be-verbatim-content column 2.

Instead, we abandon the 2-column approach and instead work in a single
column, that is vis-encoded when performing comparisons, and then
decoded to write back out to Bazel.