Use en_US encoding
The POSIX and C.UTF-8 locales seem to work different between my Mac OS
dev machine and the Ubuntu CI machines. Rather than continuing to fight
Ubuntu to convince it to emit escapes like Mac OS is doing, attempt the
opposite and try to coerce Mac OS to behave like Ubuntu and not perform
the escaping.