Create gawk versions of the vis encoding tools
Using AWK opens up new opportunities not available when using Sed.
Notably, we are able to implement these translations using a one-pass
algorithm using a hashtable to lookup replacements.

We rely on the GNU AWK implementation in particular for a few of its
extensions:
* `--characters-as-bytes` to be able to operate byte-by-byte and not be
  interfered with by locales and multibyte encoding support.
* 4-arg `split` to be gain easy access to both the content parts and the
  separators that divide them.
* NUL byte being treated as just any other character.
* `RT` to more easily implement non-newline-terminated files

Some of these are not essention, just nice to have. Others have more
verbose analogues in POSIX Sed. It may be possible to implement these
scripts without the GNU extensions, but that would likely come at the
cost of expressivity and succinctness.