chore: make root workspace buildable (#1071)
* chore: make root workspace buildable

replaces previous PR which tried to workaround this.
Remove rules_lint - the reasoning is we want this repo to be simpler since it's a common dependency for users to visit, and also our code is in Go but rules_lint has no Go linters

* fix: include_external_repositories refers to canonical name

Therefore there's a + (or earlier ~) character appended on the end of the value we compare against.
This seems undesirable.