Wrap reference to @platforms in Label()
This macro only works when it is invoked from within a repository that
has an explicit dependency on @platforms. This repo needs to be pulled
in manually nowadays, meaning that it is no longer guaranteed to be
available. By wrapping this in Label(), we force resolution using the
MODULE.bazel declared in aspect-bazel-lib itself.