refactor: remove tar toolchain (#1132)
* refactor: remove tar toolchain

It is now in tar.bzl, see https://github.com/bazel-contrib/tar.bzl/pull/37
Continued cleanup following #1123 which removed the tar rule.

Note, also found that the build was broken, fixed it up.

* fix: just check in a test fixture for zstd decompress instead of needing a real tar binary

The flags are hard to juggle between linux and macos, and we don't want to dep on tar.bzl rule just for this test